Remove unused templates for pedidos, usuarios, and zapatillas management; update products template for improved layout and functionality.
