+++ routes/main.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template
+from flask import Blueprint, render_template, redirect, url_for, session, flash, request, jsonify
 from extensions import mysql
 
 main_bp = Blueprint('main', __name__)
@@ -6,7 +6,52 @@ main_bp = Blueprint('main', __name__)
 # Ruta principal
 @main_bp.route('/')
 def index():
-    return render_template('index.html')
+    cursor = mysql.connection.cursor()
+    
+    genero = request.args.get('genero')
+    tipo = request.args.get('tipo')
+    precio = request.args.get('precio', type=float)
+    colaboracion = request.args.get('colaboracion')
+    tallas = request.args.getlist('tallas')
+
+    query = "SELECT id, nombre, descripcion, precio, imagen_url FROM zapatillas WHERE 1=1"
+    params = [] 
+
+    if genero:
+        query += " AND genero = %s"
+        params.append(genero)
+    if tipo:
+        query += " AND tipo = %s"
+        params.append(tipo)
+    if precio:
+        query += " AND precio <= %s"
+        params.append(precio)
+    if colaboracion:
+        query += " AND colaboracion LIKE %s"
+        params.append(f"%{colaboracion}%")
+    if tallas:
+        query += " AND talla IN ({})".format(','.join(['%s'] * len(tallas)))
+        params.extend(tallas)
+        
+    # obtener 3 productos aleatorios
+    if not any([genero, tipo, precio, colaboracion, tallas]):
+        query += " ORDER BY RAND() LIMIT 3"
+    else:
+        cursor.execute(query, tuple(params))
+
+    cursor.execute(query, tuple(params))
+    productos = cursor.fetchall()
+    cursor.close()
+
+    lista_productos = [{
+        'id': p['id'],
+        'nombre': p['nombre'],
+        'descripcion': p['descripcion'],
+        'precio': float(p['precio']),
+        'imagen': p['imagen_url']
+    } for p in productos]
+    
+    return render_template('index.html', productos=lista_productos)
 
 # Ruta para marcas
 @main_bp.route('/marcas')
@@ -22,11 +67,6 @@ def locales():
 @main_bp.route('/carrito')
 def carrito():
     return render_template('carrito.html')
-
-# Ruta para obtener las categorías de productos
-@main_bp.route('/category/<name>')
-def category(name):
-    return render_template('category.html', category=name)
 
 # Manejador de errores
 @main_bp.errorhandler(404)
