--- templates/admin/pedidos/index.html
@@ -1,30 +1,0 @@
-{% extends 'app.html' %}
-{% block content %}
-<h2>Pedidos</h2>
-<table class="table table-bordered">
-  <thead>
-    <tr>
-      <th>ID</th>
-      <th>Usuario</th>
-      <th>Total</th>
-      <th>Estado</th>
-      <th>Fecha</th>
-      <th>Acciones</th>
-    </tr>
-  </thead>
-  <tbody>
-    {% for p in pedidos %}
-    <tr>
-      <td>{{ p[0] }}</td>
-      <td>{{ p[1] }}</td>
-      <td>S/. {{ p[2] }}</td>
-      <td>{{ p[3] }}</td>
-      <td>{{ p[4] }}</td>
-      <td>
-        <a href="{{ url_for('admin.ver_pedido', id=p[0]) }}" class="btn btn-info btn-sm">Ver Detalles</a>
-      </td>
-    </tr>
-    {% endfor %}
-  </tbody>
-</table>
-{% endblock %}
