--- templates/admin/usuarios/editar.html
@@ -1,31 +1,0 @@
-{% extends 'app.html' %}
-{% block content %}
-<h2>Editar Usuario</h2>
-<form method="post">
-  <div class="mb-3">
-    <label>Nombre</label>
-    <input type="text" name="nombre" value="{{ usuario[1] }}" class="form-control" required>
-  </div>
-  <div class="mb-3">
-    <label>Email</label>
-    <input type="email" name="username" value="{{ usuario[2] }}" class="form-control" required>
-  </div>
-  <div class="mb-3">
-    <label>Dirección</label>
-    <input type="text" name="direccion" value="{{ usuario[3] }}" class="form-control">
-  </div>
-  <div class="mb-3">
-    <label>Teléfono</label>
-    <input type="text" name="telefono" value="{{ usuario[4] }}" class="form-control">
-  </div>
-  <div class="mb-3">
-    <label>Rol</label>
-    <select name="rol" class="form-control">
-      <option value="USER" {% if usuario[5]=='USER' %}selected{% endif %}>USER</option>
-      <option value="ADMIN" {% if usuario[5]=='ADMIN' %}selected{% endif %}>ADMIN</option>
-    </select>
-  </div>
-  <button type="submit" class="btn btn-primary">Guardar Cambios</button>
-  <a href="{{ url_for('admin.lista_usuarios') }}" class="btn btn-secondary">Cancelar</a>
-</form>
-{% endblock %}
