--- templates/admin/zapatillas/editar.html
@@ -1,9 +1,0 @@
-{% extends 'app.html' %}
-{% block content %}
-<h2>Editar Zapatilla</h2>
-<form method="post">
-  {% include 'admin/zapatillas/formulario.html' %}
-  <button type="submit" class="btn btn-primary">Actualizar</button>
-  <a href="{{ url_for('admin.lista_zapatillas') }}" class="btn btn-secondary">Cancelar</a>
-</form>
-{% endblock %}
