--- templates/admin/zapatillas/formulario.html
@@ -1,41 +1,0 @@
-<div class="mb-3">
-  <label>Nombre</label>
-  <input type="text" name="nombre" value="{{ zapatilla[1] if zapatilla else '' }}" class="form-control" required>
-</div>
-<div class="mb-3">
-  <label>Marca</label>
-  <input type="text" name="marca" value="{{ zapatilla[2] if zapatilla else '' }}" class="form-control">
-</div>
-<div class="mb-3">
-  <label>Descripción</label>
-  <textarea name="descripcion" class="form-control">{{ zapatilla[3] if zapatilla else '' }}</textarea>
-</div>
-<div class="mb-3">
-  <label>Precio</label>
-  <input type="number" step="0.01" name="precio" value="{{ zapatilla[4] if zapatilla else '' }}" class="form-control"
-    required>
-</div>
-<div class="mb-3">
-  <label>Stock</label>
-  <input type="number" name="stock" value="{{ zapatilla[5] if zapatilla else '' }}" class="form-control" required>
-</div>
-<div class="mb-3">
-  <label>Imagen URL</label>
-  <input type="text" name="imagen_url" value="{{ zapatilla[6] if zapatilla else '' }}" class="form-control">
-</div>
-<div class="mb-3">
-  <label>Talla</label>
-  <input type="text" name="talla" value="{{ zapatilla[7] if zapatilla else '' }}" class="form-control">
-</div>
-<div class="mb-3">
-  <label>Género</label>
-  <input type="text" name="genero" value="{{ zapatilla[9] if zapatilla else '' }}" class="form-control">
-</div>
-<div class="mb-3">
-  <label>Tipo</label>
-  <input type="text" name="tipo" value="{{ zapatilla[10] if zapatilla else '' }}" class="form-control">
-</div>
-<div class="mb-3">
-  <label>Colaboración</label>
-  <input type="text" name="colaboracion" value="{{ zapatilla[11] if zapatilla else '' }}" class="form-control">
-</div>
