+++ templates/carrito.html
@@ -3,7 +3,7 @@
 {% block title %}Carrito de Compras{% endblock %}
 
 {% block content %}
-<div class="max-w-4xl mx-auto p-6">
+<div class="max-w-4xl mx-auto min-h-screen p-6">
     <h1 class="text-3xl font-bold mb-6">Carrito de Compras</h1>
 
     {% if productos %}
