--- templates/carritoadm.html
@@ -1,109 +1,0 @@
-{% extends 'app.html' %}
-
-{% block title %}Administrar Carrito{% endblock %}
-
-{% block content %}
-<div class="max-w-5xl mx-auto p-6">
-    <h1 class="text-2xl font-bold mb-4">Carrito de Compras</h1>
-
-    <!-- Botón para agregar -->
-    <div class="flex justify-end mb-4">
-        <button onclick="abrirPanelAgregar()" class="bg-blue-600 text-white px-4 py-2 rounded hover:bg-blue-700">
-            Agregar al Carrito
-        </button>
-    </div>
-
-    <!-- Tabla -->
-    <div class="overflow-x-auto">
-        <table class="w-full table-auto border">
-            <thead class="bg-gray-100 text-left">
-                <tr>
-                    <th class="p-2">ID</th>
-                    <th class="p-2">Usuario</th>
-                    <th class="p-2">Zapatilla</th>
-                    <th class="p-2">Cantidad</th>
-                    <th class="p-2">Agregado En</th>
-                    <th class="p-2">Acciones</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for item in carrito %}
-                <tr class="border-t">
-                    <td class="p-2">{{ item.id }}</td>
-                    <td class="p-2">{{ item.nombre_usuario }} (ID {{ item.usuario_id }})</td>
-                    <td class="p-2">{{ item.nombre_zapatilla }} (ID {{ item.zapatilla_id }})</td>
-                    <td class="p-2">{{ item.cantidad }}</td>
-                    <td class="p-2">{{ item.agregado_en }}</td>
-                    <td class="p-2 space-x-2">
-                        <button onclick="abrirPanelEditar({{ item.id }}, {{ item.usuario_id }}, {{ item.zapatilla_id }}, {{ item.cantidad }})"
-                                class="bg-yellow-400 text-white px-2 py-1 rounded hover:bg-yellow-500">Editar</button>
-                        <form action="{{ url_for('carritoadm.eliminar_carrito', id=item.id) }}" method="POST" style="display:inline;">
-                            <button type="submit" class="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                                    onclick="return confirm('¿Eliminar este ítem del carrito?')">Eliminar</button>
-                        </form>
-                    </td>
-                </tr>
-                {% else %}
-                <tr>
-                    <td colspan="6" class="text-center text-gray-500 p-4">El carrito está vacío.</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-</div>
-
-<!-- Panel lateral -->
-<div id="panelCarrito" class="fixed top-0 right-0 w-full max-w-md h-full bg-white shadow-lg transform translate-x-full transition-transform duration-300 z-50 overflow-y-auto">
-    <div class="p-6">
-        <div class="flex justify-between items-center mb-4">
-            <h2 id="tituloPanel" class="text-xl font-bold">Agregar al Carrito</h2>
-            <button onclick="cerrarPanel()" class="text-gray-500 text-2xl hover:text-black">&times;</button>
-        </div>
-        <form id="formCarrito" method="POST">
-            <input type="hidden" name="id" id="input_id">
-            <div class="space-y-4">
-                <div>
-                    <label class="block text-sm font-medium">ID Usuario</label>
-                    <input type="number" name="usuario_id" id="input_usuario" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">ID Zapatilla</label>
-                    <input type="number" name="zapatilla_id" id="input_zapatilla" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Cantidad</label>
-                    <input type="number" name="cantidad" id="input_cantidad" min="1" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <button type="submit" class="bg-green-600 text-white px-4 py-2 rounded hover:bg-green-700">Guardar</button>
-            </div>
-        </form>
-    </div>
-</div>
-
-<script>
-    const panel = document.getElementById('panelCarrito');
-    const form = document.getElementById('formCarrito');
-    const titulo = document.getElementById('tituloPanel');
-
-    function abrirPanelAgregar() {
-        titulo.textContent = "Agregar al Carrito";
-        form.action = "{{ url_for('carritoadm.agregar_carrito') }}";
-        form.reset();
-        panel.classList.remove('translate-x-full');
-    }
-
-    function abrirPanelEditar(id, usuario_id, zapatilla_id, cantidad) {
-        titulo.textContent = "Editar Ítem del Carrito";
-        form.action = `/carritoadm/editar/${id}`;
-        document.getElementById('input_usuario').value = usuario_id;
-        document.getElementById('input_zapatilla').value = zapatilla_id;
-        document.getElementById('input_cantidad').value = cantidad;
-        panel.classList.remove('translate-x-full');
-    }
-
-    function cerrarPanel() {
-        panel.classList.add('translate-x-full');
-    }
-</script>
-{% endblock %}
