--- templates/pedidosadm.html
@@ -1,116 +1,0 @@
-{% extends 'app.html' %}
-{% block title %}Gestión de Pedidos{% endblock %}
-
-{% block content %}
-<div class="max-w-6xl mx-auto p-6">
-    <h1 class="text-3xl font-bold mb-6 text-center">Gestión de Pedidos</h1>
-
-    <!-- Botón agregar -->
-    <div class="flex justify-end mb-4">
-        <button onclick="abrirPanelAgregar()" class="bg-blue-600 text-white px-4 py-2 rounded hover:bg-blue-700">
-            Nuevo Pedido
-        </button>
-    </div>
-
-    <div class="overflow-x-auto rounded-lg shadow">
-        <table class="min-w-full bg-white text-sm text-left text-gray-700">
-            <thead class="bg-gray-100 uppercase text-xs font-semibold border-b border-gray-200">
-                <tr>
-                    <th class="px-4 py-3">ID</th>
-                    <th class="px-4 py-3">Usuario</th>
-                    <th class="px-4 py-3">Total</th>
-                    <th class="px-4 py-3">Estado</th>
-                    <th class="px-4 py-3">Dirección</th>
-                    <th class="px-4 py-3">Fecha Pedido</th>
-                    <th class="px-4 py-3 text-center">Acciones</th>
-                </tr>
-            </thead>
-            <tbody class="divide-y divide-gray-200">
-                {% for pedido in pedidos %}
-                <tr class="hover:bg-gray-50">
-                    <td class="px-4 py-2">{{ pedido.id }}</td>
-                    <td class="px-4 py-2">{{ pedido.usuario_id }}</td>
-                    <td class="px-4 py-2">${{ pedido.total }}</td>
-                    <td class="px-4 py-2">{{ pedido.estado }}</td>
-                    <td class="px-4 py-2">{{ pedido.direccion }}</td>
-                    <td class="px-4 py-2">{{ pedido.fecha_pedido }}</td>
-                    <td class="px-4 py-2 text-center space-x-2">
-                        <button onclick="abrirPanelEditar({{ pedido.id }}, '{{ pedido.usuario_id }}', '{{ pedido.total }}', '{{ pedido.estado }}', '{{ pedido.direccion }}', '{{ pedido.fecha_pedido }}')" class="bg-yellow-500 hover:bg-yellow-600 text-white px-3 py-1 rounded text-xs shadow">Editar</button>
-                        <a href="{{ url_for('pedidosadm.eliminar_pedido', id=pedido.id) }}" class="bg-red-500 hover:bg-red-600 text-white px-3 py-1 rounded text-xs shadow" onclick="return confirm('¿Eliminar este pedido?');">Eliminar</a>
-                    </td>
-                </tr>
-                {% else %}
-                <tr>
-                    <td colspan="7" class="text-center text-gray-500 py-4">No hay pedidos registrados.</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-</div>
-
-<!-- Panel lateral -->
-<div id="panelPedido" class="fixed top-0 right-0 w-full max-w-md h-full bg-white shadow-lg transform translate-x-full transition-transform duration-300 z-50 overflow-y-auto">
-    <div class="p-6">
-        <div class="flex justify-between items-center mb-4">
-            <h2 id="tituloPanel" class="text-xl font-bold">Agregar Pedido</h2>
-            <button onclick="cerrarPanel()" class="text-gray-500 text-2xl hover:text-black">&times;</button>
-        </div>
-        <form id="formPedido" method="POST">
-            <input type="hidden" name="id" id="input_id">
-            <div class="space-y-4">
-                <div>
-                    <label class="block text-sm font-medium">Usuario ID</label>
-                    <input type="number" name="usuario_id" id="input_usuario_id" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Total</label>
-                    <input type="number" name="total" step="0.01" id="input_total" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Estado</label>
-                    <input type="text" name="estado" id="input_estado" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Dirección</label>
-                    <input type="text" name="direccion" id="input_direccion" class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Fecha Pedido</label>
-                    <input type="date" name="fecha_pedido" id="input_fecha_pedido" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <button type="submit" class="bg-green-600 text-white px-4 py-2 rounded hover:bg-green-700">Guardar</button>
-            </div>
-        </form>
-    </div>
-</div>
-
-<script>
-    const panel = document.getElementById('panelPedido');
-    const titulo = document.getElementById('tituloPanel');
-    const form = document.getElementById('formPedido');
-
-    function abrirPanelAgregar() {
-        titulo.textContent = "Agregar Pedido";
-        form.action = "{{ url_for('pedidosadm.agregar_pedido') }}";
-        form.reset();
-        panel.classList.remove('translate-x-full');
-    }
-
-    function abrirPanelEditar(id, usuario_id, total, estado, direccion, fecha_pedido) {
-        titulo.textContent = "Editar Pedido";
-        form.action = `/pedidosadm/editar/${id}`;
-        document.getElementById('input_id').value = id;
-        document.getElementById('input_usuario_id').value = usuario_id;
-        document.getElementById('input_total').value = total;
-        document.getElementById('input_estado').value = estado;
-        document.getElementById('input_direccion').value = direccion;
-        document.getElementById('input_fecha_pedido').value = fecha_pedido;
-        panel.classList.remove('translate-x-full');
-    }
-
-    function cerrarPanel() {
-        panel.classList.add('translate-x-full');
-    }
-</script>
-{% endblock %}
