--- templates/usuariosadm.html
@@ -1,126 +1,0 @@
-{% extends 'app.html' %}
-
-{% block title %}Administrar Usuarios{% endblock %}
-
-{% block content %}
-<div class="max-w-6xl mx-auto px-4 py-8">
-    <h1 class="text-3xl font-bold mb-6 text-center">Gestión de Usuarios</h1>
-
-    <!-- Botón para añadir nuevo usuario -->
-    <div class="flex justify-end mb-4">
-        <button onclick="abrirPanelAgregar()" class="bg-blue-600 text-white px-4 py-2 rounded hover:bg-blue-700">
-            Nuevo Usuario
-        </button>
-    </div>
-
-    <!-- Tabla de usuarios -->
-    <div class="bg-white shadow-md rounded-lg overflow-x-auto">
-        <table class="min-w-full divide-y divide-gray-200">
-            <thead class="bg-gray-100">
-                <tr>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase">ID</th>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase">Nombre</th>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase">Usuario</th>
-                    <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase">Rol</th>
-                    <th class="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase">Acciones</th>
-                </tr>
-            </thead>
-            <tbody class="bg-white divide-y divide-gray-200">
-                {% for usuario in usuarios %}
-                <tr>
-                    <td class="px-6 py-4 whitespace-nowrap">{{ usuario.id }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap">{{ usuario.nombre }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap">{{ usuario.username }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap">{{ usuario.rol }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap text-right text-sm font-medium space-x-2">
-                        <button onclick="abrirPanelEditar({{ usuario.id }}, '{{ usuario.nombre }}', '{{ usuario.username }}', '{{ usuario.direccion }}', '{{ usuario.telefono }}', '{{ usuario.rol }}')" class="text-yellow-500 hover:text-yellow-600">Editar</button>
-                        <form action="{{ url_for('usuariosadm.eliminar_usuario', id=usuario.id) }}" method="POST" style="display:inline;">
-                            <button type="submit" class="text-red-500 hover:text-red-600" onclick="return confirm('¿Estás seguro de que deseas eliminar este usuario?');">Eliminar</button>
-                        </form>
-                    </td>
-                </tr>
-                {% else %}
-                <tr>
-                    <td colspan="5" class="px-6 py-4 text-center text-gray-500">No hay usuarios registrados.</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-</div>
-
-<!-- Panel lateral para agregar/editar usuario -->
-<div id="panelUsuario" class="fixed top-0 right-0 w-full max-w-md h-full bg-white shadow-lg transform translate-x-full transition-transform duration-300 z-50 overflow-y-auto">
-    <div class="p-6">
-        <div class="flex justify-between items-center mb-4">
-            <h2 id="tituloPanel" class="text-xl font-bold">Agregar Usuario</h2>
-            <button onclick="cerrarPanel()" class="text-gray-500 text-2xl hover:text-black">&times;</button>
-        </div>
-        <form id="formUsuario" method="POST">
-            <input type="hidden" name="id" id="input_id">
-            <div class="space-y-4">
-                <div>
-                    <label class="block text-sm font-medium">Nombre</label>
-                    <input type="text" name="nombre" id="input_nombre" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Usuario</label>
-                    <input type="text" name="username" id="input_username" required class="w-full border px-3 py-2 rounded">
-                </div>
-                <div id="campoPassword">
-                    <label class="block text-sm font-medium">Contraseña</label>
-                    <input type="password" name="password" id="input_password" class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Dirección</label>
-                    <input type="text" name="direccion" id="input_direccion" class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Teléfono</label>
-                    <input type="text" name="telefono" id="input_telefono" class="w-full border px-3 py-2 rounded">
-                </div>
-                <div>
-                    <label class="block text-sm font-medium">Rol</label>
-                    <select name="rol" id="input_rol" required class="w-full border px-3 py-2 rounded">
-                        <option value="USUARIO">USUARIO</option>
-                        <option value="ADMIN">ADMIN</option>
-                    </select>
-                </div>
-                <button type="submit" class="bg-green-600 text-white px-4 py-2 rounded hover:bg-green-700">Guardar</button>
-            </div>
-        </form>
-    </div>
-</div>
-
-<script>
-    const panel = document.getElementById('panelUsuario');
-    const titulo = document.getElementById('tituloPanel');
-    const form = document.getElementById('formUsuario');
-    const campoPassword = document.getElementById('campoPassword');
-
-    function abrirPanelAgregar() {
-        titulo.textContent = "Agregar Usuario";
-        form.action = "{{ url_for('usuariosadm.agregar_usuario') }}";
-        form.reset();
-        campoPassword.style.display = 'block';
-        panel.classList.remove('translate-x-full');
-    }
-
-    function abrirPanelEditar(id, nombre, username, direccion, telefono, rol) {
-        titulo.textContent = "Editar Usuario";
-        form.action = `/usuariosadm/editar/${id}`;
-        document.getElementById('input_id').value = id;
-        document.getElementById('input_nombre').value = nombre;
-        document.getElementById('input_username').value = username;
-        document.getElementById('input_direccion').value = direccion;
-        document.getElementById('input_telefono').value = telefono;
-        document.getElementById('input_rol').value = rol;
-        campoPassword.style.display = 'none'; // No se edita contraseña
-        panel.classList.remove('translate-x-full');
-    }
-
-    function cerrarPanel() {
-        panel.classList.add('translate-x-full');
-    }
-</script>
-{% endblock %}
